Split `EditableVnd` to its own file.
