+++ lib/flutter_vnd_keyboard.dart
@@ -1,3 +1,4 @@
+export 'src/editable_vnd.dart';
 export 'src/keyboard_key.dart';
 export 'src/vnd_bottom_sheet.dart';
 export 'src/vnd_editing_controller.dart';
